requirements.txt
Django==4.2.7
Pillow==10.1.0
python-dotenv==1.0.0 